fixing sequelize, routes, and everything
