+++ .eslintignore
@@ -1,2 +1,3 @@
 models/index.js
 node_modules
+config/config.json
