+++ models/losers.js
@@ -1,16 +1,18 @@
 var Sequelize = require("sequelize");
-var sequelize = require("../config/connection.js");
 
-var Losers = sequelize.define("loser", {
+var Losers = sequelize.define(
+  "loser",
+  {
+    missionName: Sequelize.STRING,
 
-  missionName: Sequelize.STRING,
+    days: Sequelize.INTEGER,
 
-  days: Sequelize.INTEGER,
-
-  epitaph: Sequelize.INTEGER,
-}, {
-  timestamps: true
-});
+    epitaph: Sequelize.INTEGER
+  },
+  {
+    timestamps: true
+  }
+);
 
 // Makes the Character Model available for other files (will also create a table)
 module.exports = Losers;
