+++ package.json
@@ -16,7 +16,7 @@
     "express": "^4.16.3",
     "express-handlebars": "^3.0.0",
     "mysql2": "^1.5.3",
-    "sequelize": "^4.38.0"
+    "sequelize": "^4.38.1"
   },
   "devDependencies": {
     "chai": "^4.1.2",
