+++ routes/htmlRoutes.js
@@ -1,14 +1,9 @@
-var db = require("../models");
+var db = require("../config/connection.js");
 
 module.exports = function(app) {
   // Load index page
   app.get("/", function(req, res) {
-    db.Example.findAll({}).then(function(dbExamples) {
-      res.render("index", {
-        msg: "Welcome!",
-        examples: dbExamples
-      });
-    });
+    res.render("index");
   });
 
   app.get("/play", function(req, res) {
@@ -22,19 +17,19 @@ module.exports = function(app) {
     res.render("win");
   });
 
-  // Load example page and pass in an example by id
-  app.get("/example/:id", function(req, res) {
-    db.Example.findOne({ where: { id: req.params.id } }).then(function(
-      dbExample
-    ) {
-      res.render("example", {
-        example: dbExample
-      });
-    });
-  });
+//   // Load example page and pass in an example by id
+//   app.get("/example/:id", function(req, res) {
+//     db.Example.findOne({ where: { id: req.params.id } }).then(function(
+//       dbExample
+//     ) {
+//       res.render("example", {
+//         example: dbExample
+//       });
+//     });
+//   });
 
-  // Render 404 page for any unmatched routes
-  app.get("*", function(req, res) {
-    res.render("404");
-  });
-};
+//   // Render 404 page for any unmatched routes
+//   app.get("*", function(req, res) {
+//     res.render("404");
+//   });
+// };
