+++ server.js
@@ -3,7 +3,7 @@ var express = require("express");
 var bodyParser = require("body-parser");
 var exphbs = require("express-handlebars");
 
-var db = require("./models");
+var db = require("./config/connection.js");
 
 var app = express();
 var PORT = process.env.PORT || 3000;
